feat: new help commands related to formatters
