+++ pkg/commands/help.go
@@ -1,10 +1,6 @@
 package commands
 
 import (
-	"encoding/json"
-	"fmt"
-	"maps"
-	"slices"
 	"strings"
 	"unicode"
 	"unicode/utf8"
@@ -12,22 +8,9 @@ import (
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 
-	"github.com/golangci/golangci-lint/pkg/config"
-	"github.com/golangci/golangci-lint/pkg/lint/linter"
 	"github.com/golangci/golangci-lint/pkg/lint/lintersdb"
 	"github.com/golangci/golangci-lint/pkg/logutils"
 )
-
-type linterHelp struct {
-	Name        string   `json:"name"`
-	Desc        string   `json:"description"`
-	Groups      []string `json:"groups"`
-	Fast        bool     `json:"fast"`
-	AutoFix     bool     `json:"autoFix"`
-	Deprecated  bool     `json:"deprecated"`
-	Since       string   `json:"since"`
-	OriginalURL string   `json:"originalURL,omitempty"`
-}
 
 type helpOptions struct {
 	JSON bool
@@ -60,11 +43,22 @@ func newHelpCommand(logger logutils.Log) *helpCommand {
 		Short:             "Help about linters",
 		Args:              cobra.NoArgs,
 		ValidArgsFunction: cobra.NoFileCompletions,
-		RunE:              c.execute,
-		PreRunE:           c.preRunE,
+		RunE:              c.lintersExecute,
+		PreRunE:           c.lintersPreRunE,
 	}
 
 	helpCmd.AddCommand(lintersCmd)
+
+	formattersCmd := &cobra.Command{
+		Use:               "formatters",
+		Short:             "Help about formatters",
+		Args:              cobra.NoArgs,
+		ValidArgsFunction: cobra.NoFileCompletions,
+		RunE:              c.formattersExecute,
+		PreRunE:           c.formattersPreRunE,
+	}
+
+	helpCmd.AddCommand(formattersCmd)
 
 	fs := lintersCmd.Flags()
 	fs.SortFlags = false // sort them as they are defined here
@@ -74,122 +68,6 @@ func newHelpCommand(logger logutils.Log) *helpCommand {
 	c.cmd = helpCmd
 
 	return c
-}
-
-func (c *helpCommand) preRunE(_ *cobra.Command, _ []string) error {
-	// The command doesn't depend on the real configuration.
-	// It just needs the list of all plugins and all presets.
-	dbManager, err := lintersdb.NewManager(c.log.Child(logutils.DebugKeyLintersDB), config.NewDefault(), lintersdb.NewLinterBuilder())
-	if err != nil {
-		return err
-	}
-
-	c.dbManager = dbManager
-
-	return nil
-}
-
-func (c *helpCommand) execute(_ *cobra.Command, _ []string) error {
-	if c.opts.JSON {
-		return c.printJSON()
-	}
-
-	c.print()
-
-	return nil
-}
-
-func (c *helpCommand) printJSON() error {
-	var linters []linterHelp
-
-	for _, lc := range c.dbManager.GetAllSupportedLinterConfigs() {
-		if lc.Internal {
-			continue
-		}
-
-		groups := []string{config.GroupAll}
-
-		if !lc.IsSlowLinter() {
-			groups = append(groups, config.GroupFast)
-		}
-
-		linters = append(linters, linterHelp{
-			Name:        lc.Name(),
-			Desc:        formatDescription(lc.Linter.Desc()),
-			Groups:      slices.Concat(groups, slices.Collect(maps.Keys(lc.Groups))),
-			Fast:        !lc.IsSlowLinter(),
-			AutoFix:     lc.CanAutoFix,
-			Deprecated:  lc.IsDeprecated(),
-			Since:       lc.Since,
-			OriginalURL: lc.OriginalURL,
-		})
-	}
-
-	return json.NewEncoder(c.cmd.OutOrStdout()).Encode(linters)
-}
-
-func (c *helpCommand) print() {
-	var enabledLCs, disabledLCs []*linter.Config
-	for _, lc := range c.dbManager.GetAllSupportedLinterConfigs() {
-		if lc.Internal {
-			continue
-		}
-
-		if lc.FromGroup(config.GroupStandard) {
-			enabledLCs = append(enabledLCs, lc)
-		} else {
-			disabledLCs = append(disabledLCs, lc)
-		}
-	}
-
-	color.Green("Enabled by default linters:\n")
-	printLinters(enabledLCs)
-
-	color.Red("\nDisabled by default linters:\n")
-	printLinters(disabledLCs)
-}
-
-func printLinters(lcs []*linter.Config) {
-	slices.SortFunc(lcs, func(a, b *linter.Config) int {
-		if a.IsDeprecated() && b.IsDeprecated() {
-			return strings.Compare(a.Name(), b.Name())
-		}
-
-		if a.IsDeprecated() {
-			return 1
-		}
-
-		if b.IsDeprecated() {
-			return -1
-		}
-
-		return strings.Compare(a.Name(), b.Name())
-	})
-
-	for _, lc := range lcs {
-		desc := formatDescription(lc.Linter.Desc())
-
-		deprecatedMark := ""
-		if lc.IsDeprecated() {
-			deprecatedMark = " [" + color.RedString("deprecated") + "]"
-		}
-
-		var capabilities []string
-		if !lc.IsSlowLinter() {
-			capabilities = append(capabilities, color.BlueString("fast"))
-		}
-		if lc.CanAutoFix {
-			capabilities = append(capabilities, color.GreenString("auto-fix"))
-		}
-
-		var capability string
-		if capabilities != nil {
-			capability = " [" + strings.Join(capabilities, ", ") + "]"
-		}
-
-		_, _ = fmt.Fprintf(logutils.StdOut, "%s%s: %s%s\n",
-			color.YellowString(lc.Name()), deprecatedMark, desc, capability)
-	}
 }
 
 func formatDescription(desc string) string {
